requirements.txt
nltk==3.8.1
langchain==0.1.16
langchain-openai==0.1.3
langchain_google_genai==1.0.2
transformers==4.41.2
accelerate==0.29.3
python-dotenv==1.0.1
gradio==4.29.0
black==24.4.0
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2
chardet==5.2.0
sentencepiece==0.2.0
evaluate==0.4.2
rouge_score==0.1.2
pytest==8.2.1
seaborn==0.13.2
# xformers 
trl<0.9.0
peft 
accelerate 
bitsandbytes
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
jupyter 
ipywidgets